Run deprecated import check before imports from other tests
